Add ITKMontage as an external project and add option to build ITK v5.0.0
ITKMontage is used by the plugin `MultiscaleFusion`. It is not yet available
as a remote module inside ITK, and therefore has to be built separately.
An option, to select ITKv4.13.0 or ITKV5.0.0 has been added. By default v4.13.0
is built, but if v5.0.0 is selected ITKMontage remote module is built.